% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakeFunctions.R
\name{cha}
\alias{cha}
\title{cha: convex hull area}
\usage{
cha(x, y)
}
\arguments{
\item{x}{dataframe of longitudes}

\item{y}{dataframe of latitudes}
}
\value{
area of the polygon
}
\description{
cha: convex hull area
}
\examples{
Usage example
}
