% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakeFunctions.R
\name{circle}
\alias{circle}
\title{circle: calculate a set of points that lie on a circle with a given center and diameter}
\usage{
circle(center = c(0, 0), diameter = 1, npoints = 100)
}
\arguments{
\item{center}{a vector containing the center coordinates of the circle}

\item{diameter}{the diameter of the circle}

\item{npoints}{the number of points to generate for the circle, default=100}
}
\value{
a dataframe containing the specified number of points that lie on the circle
}
\description{
circle: calculate a set of points that lie on a circle with a given center and diameter
}
\examples{
Usage example
}
