% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakeFunctions.R
\name{circle_intersection}
\alias{circle_intersection}
\title{circle_intersection: calculate the intersection of two circles}
\usage{
circle_intersection(center1 = c(x1, y1), radius1, center2 = c(x2, y2), radius2)
}
\arguments{
\item{center1}{a vector containing the center coordinates of the first circle}

\item{radius1}{the radius of the first circle}

\item{center2}{a vector containing the center coordinates of the second circle}

\item{radius2}{the radius of the second circle}
}
\value{
a matrix containing either the ordered pairs of the 2 intersection points or NAs if no intersection is detected
}
\description{
circle_intersection: calculate the intersection of two circles
}
\examples{
Usage example
}
