% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakeFunctions.R
\name{plot3compEqData}
\alias{plot3compEqData}
\title{plot3compEqData: plot all components of 3-component (Z, N, E) seismic data in one window with labels}
\usage{
plot3compEqData(
  data.df,
  total.time = 27,
  plotTitle = "plot",
  mag = "mag",
  starttime = "starttime",
  sta = "sta"
)
}
\arguments{
\item{data.df}{dataframe containing sesimic data for 3 components}

\item{total.time}{time interval for the data, default 27s for LEN-DB data}

\item{plotTitle}{title of the plot}

\item{mag}{attribute magnitude}

\item{starttime}{attribute starttime}

\item{sta}{station code}
}
\value{
ggplot of the seismic data
}
\description{
plot3compEqData: plot all components of 3-component (Z, N, E) seismic data in one window with labels
}
\examples{
Usage example
}
